requirements.txt
Django
djangorestframework
django-cors-headers
whitenoise
gunicorn
inflection
PyYAML
uritemplate
psycopg2-binary